+++ NEWS.md
@@ -1,6 +1,7 @@
 # officer.pptx 0.1.1
 
-* `slide_preview` to plot a slide, `slide_to_png` to save a slide to file
+* `slide_preview` to plot a slide, `slide_to_png` to save a slide to file. Experimental version `slide_preview_2` 
+  leveraging the `doconv` package, but with some problems (#5).
 * `add_slide`: gains dynamic dots (passed on to `phs_with`) and arg `.phs_with` to create slide and fill placeholders in one step.
 * `phs_with` gains dynamic dots.
 * `xml_add_image_at_shape`: fix vectorization. Only works for non placeholder shapes.